pieniä korjauksia

+++ controllers/pelaajat.js
@@ -11,6 +11,25 @@ pelaajatRouter.get('/', (req, res) => {
     )
 })
 
+// GET pelaajat/{id}
+pelaajatRouter.get('/:id', (request, response, next) => {
+    const id = request.params.id 
+    if (!ObjectId.isValid(id)) return response.status(404).end()
+    try {
+        Pelaaja.findById(id).then(pelaaja => {
+            if (pelaaja) {
+                response.json(pelaaja)
+            } else {
+                response.status(404).end()
+            }
+        })
+    } catch (error) {
+        next(error)   
+    }
+
+})
+
+
 //POST pelaajat
 pelaajatRouter.post('/', async (request, response, next) => {
     const body = request.body
@@ -42,23 +61,38 @@ pelaajatRouter.post('/', async (request, response, next) => {
 
     }
 })
-// GET pelaajat/{id}
-pelaajatRouter.get('/:id', (request, response, next) => {
-    const id = request.params.id 
-    if (!ObjectId.isValid(id)) return response.status(404).end()
+
+//PUT pelaajat/{id}
+pelaajatRouter.put('/', async (request, response, next) => {
+    const body = request.body;
+    const id = body.id;
+    const ip = request.headers.ip
+    const o_id = new ObjectId(id);
+    
     try {
-        Pelaaja.findById(id).then(pelaaja => {
-            if (pelaaja) {
-                response.json(pelaaja)
-            } else {
-                response.status(404).end()
+        if (body.nimi) {
+            const sameNameDiffId = await Pelaaja.find({ nimi: body.nimi, ip: ip }).where({ _id: { $ne: o_id } })
+            console.log(sameNameDiffId)
+            if (sameNameDiffId.length > 0) {
+                return response.status(400).json({
+                    error: 'A player with the same name already exists.'
+                })
             }
-        })
-    } catch (error) {
-        next(error)   
+        }
+        const pelaaja = {
+            nimi: body.nimi,
+            lipukkeet: body.lipukkeet
+        }
+        const updatedPelaaja = await Pelaaja.findByIdAndUpdate(id, pelaaja, { new: true, runValidators: true, context: 'query' })
+        if (!updatedPelaaja) return response.status(404).end()
+        response.json(updatedPelaaja);
+    }
+    catch (error) {
+        next(error)
     }
 
 })
+
 
 //PUT pelaajat/{id}
 pelaajatRouter.put('/:id', async (request, response, next) => {
@@ -95,6 +129,7 @@ pelaajatRouter.put('/:id', async (request, response, next) => {
 //DELETE pelaajat/{id}
 pelaajatRouter.delete('/:id', async (request, response, next) => {
     const id = request.params.id
+    if (!ObjectId.isValid(id)) return response.status(204).end()
     try {
         await Pelaaja.findByIdAndDelete(id)
         response.status(204).end()
